Introduce noexcept use
