+++ src/cs_top_is_question.h
@@ -1,0 +1,20 @@
+/**
+ * @file cs_top_is_question.h
+ *
+ * @author  Guy Maurel
+ * @license GPL v2+
+ * extract from combine.cpp
+ */
+
+#ifndef CS_TOP_IS_QUESTION_H_INCLUDED
+#define CS_TOP_IS_QUESTION_H_INCLUDED
+
+//#include "chunk_list.h"
+#include "ChunkStack.h"
+//#include "uncrustify_types.h"
+
+
+bool cs_top_is_question(ChunkStack &cs, size_t level);
+
+
+#endif /* CS_TOP_IS_QUESTION_H_INCLUDED */
