+++ tests/cpp.test
@@ -10,6 +10,7 @@
 30001  sp_angle_colon.cfg                   cpp/alt_tokens.cpp
 30002  ben_003.cfg                          cpp/constructor.cpp
 30003  ben_004.cfg                          cpp/strings.cpp
+
 30010  ben_005.cfg                          cpp/class.h
 30011  ben_006.cfg                          cpp/misc.cpp
 30012  avalon2.cfg                          cpp/misc2.cpp
@@ -53,15 +54,14 @@
 30047  nl_func_paren_empty.cfg              cpp/nl_func_paren_empty.cpp
 30048  nl_func_def_paren_empty-r.cfg        cpp/nl_func_def_paren_empty.cpp
 30049  nl_func_call_paren-f.cfg             cpp/nl_func_call_paren.cpp
-30050  nl_namespace-r.cfg                   cpp/nl-namespace.h
 
+30050  nl_namespace-r.cfg                   cpp/nl-namespace.h
 30051  nl_namespace-a.cfg                   cpp/nl-namespace.h
 30052  nl_try-a.cfg                         cpp/try-catch-nl.cpp
 30053  sp_catch_brace.cfg                   cpp/exception.cpp
 30054  Issue_2091.cfg                       cpp/Issue_2091.cpp
 30055  nl_after_func_body.cfg               cpp/nl_func.cpp
 30056  nl_after_func_body-2.cfg             cpp/nl_func.cpp
-
 30057  nl_inside_namespace_1.cfg            cpp/nl_inside_namespace.cpp
 30058  nl_inside_namespace_2.cfg            cpp/nl_inside_namespace.cpp
 
@@ -72,10 +72,10 @@
 30064  class-colon-pos-sol-add.cfg          cpp/class-init.cpp
 30065  class-colon-pos-sol.cfg              cpp/Example.h
 30066  class-on-colon-indent.cfg            cpp/class-init.cpp
-
 30067  nl_func_class_scope-a.cfg            cpp/nl_func_type_name.cpp
 30068  nl_func_class_scope-a.cfg            cpp/nl_func_scope_name.cpp
 30069  nl_func_class_scope-a.cfg            cpp/class-implementation.cpp
+
 30070  nl_func_scope_name-a.cfg             cpp/nl_func_scope_name.cpp
 30071  lineEndings-to-Win.cfg               cpp/lineEndings-Mac.cpp
 30072  lineEndings-to-Unix.cfg              cpp/lineEndings-Win.cpp
@@ -96,8 +96,8 @@
 30095  ben_017.cfg                          cpp/bug_485.cpp
 30096  bug_1854.cfg                         cpp/bug_1854.cpp
 30097  align-1.cfg                          cpp/issue_1946.cpp
-
 30099  sp_arith-f.cfg                       cpp/bug_1127.cpp
+
 30100  nl_template_class-force.cfg          cpp/templates.cpp
 30101  nl_template_class-remove.cfg         cpp/templates.cpp
 30102  template_sp-force.cfg                cpp/templates.cpp
@@ -125,13 +125,14 @@
 30207  indent_class-t_columns-4.cfg         cpp/cmt_indent_pp.h
 30208  bug_1108.cfg                         cpp/bug_1108.cpp
 30209  empty.cfg                            cpp/bug_1134.cpp
+
 30210  empty.cfg                            cpp/bug_1338.cpp
 30211  space_indent_class-t_columns-4.cfg   cpp/indent_comment_align_thresh.cpp
 30212  indent_comment_align_thresh_2.cfg    cpp/indent_comment_align_thresh.cpp
 30213  align_right_comment.cfg              cpp/align_right_comment.cpp
 30214  issue_1887.cfg                       cpp/align_across_braces.cpp
 30215  empty.cfg                            cpp/Issue_2099.cpp
-
+30216  Issue_2302.cfg                       cpp/Issue_2302.cpp
 30217  empty.cfg                            cpp/2138.cpp
 
 30220  bug_1340.cfg                         cpp/bug_1340.cpp
