+++ tests/expected/cpp/30216-Issue_2302.cpp
@@ -1,0 +1,7 @@
+template<class T>
+class Foo<T>::Baz {
+Baz() noexcept
+	: i(0)
+{
+}
+};
