+++ tests/input/cpp/Issue_2302.cpp
@@ -1,0 +1,6 @@
+template<class T>
+class Foo<T>::Baz {
+  Baz() noexcept
+    : i(0)
+  {}
+};
